back-end/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-graphql-jwt==0.4.0
django-image-optimizer==1.0.3
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-python==0.0.1
graphql-relay==3.2.0
idna==3.7
pillow==10.3.0
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyparsing==2.0.3
python-decouple==3.8
python-dotenv==1.0.1
python-resize-image==1.1.20
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tinify==1.6.0
urllib3==2.2.1
